add websocket and switch to faster-whisper

+++ src/ai/stt.py
@@ -1,22 +1,36 @@
 from concurrent.futures import ThreadPoolExecutor
-import whisper
+from faster_whisper import WhisperModel
 import tempfile
 import asyncio
 import os
-from fastapi import APIRouter, UploadFile, File
+from fastapi import APIRouter, WebSocket, WebSocketDisconnect
 
-model = whisper.load_model("medium")
 
-executor = ThreadPoolExecutor(max_workers=3)
+
+model = WhisperModel(
+    "small",
+    device="cpu",
+    compute_type="int8"
+)
+
+executor = ThreadPoolExecutor(max_workers=2)
 
 router = APIRouter()
 
 def transcribe_sync(path: str):
-    return model.transcribe(
+    segments, info = model.transcribe(
         path,
-        language=None,
-        condition_on_previous_text=True
+        beam_size=2,
+        vad_filter=True
     )
+    
+    text = " ".join(seg.text for seg in segments)
+
+    return {
+        "text": text,
+        "language": info.language,
+        "duration": info.duration
+    }
 
 async def transcribe_async(path: str):
     loop = asyncio.get_event_loop()
@@ -26,20 +40,35 @@ async def transcribe_async(path: str):
         lambda: transcribe_sync(path)
     )
 
-@router.post("/transcribe")
-async def stt(file: UploadFile = File(...)):
+@router.websocket("/ws/stt")
+async def websocket_stt(ws: WebSocket):
+    await ws.accept()
 
-    with tempfile.NamedTemporaryFile(delete=False, suffix=".wav") as tmp:
-        path = tmp.name
-        try:
-            while chunk := await file.read(1024 * 1024):
-                tmp.write(chunk)
+    tmp = tempfile.NamedTemporaryFile(delete=False, suffix=".wav")
+    path = tmp.name
+    try:
+        while True:
+            data = await ws.receive_bytes()
+
+            # !!! Тут потрібно щоб фронт надсилав __start__ і __end__ для того щоб ШІ розуміла !!!
             
-            tmp.close()
-        
-            result = await transcribe_async(path)
             
-            return {"result": result}
-        finally:
+            if data == b"__end__":
+
+                text = await transcribe_async(path)
+
+                await ws.send_json(text)
+
+                await ws.close()
+                tmp.close()
+                break
+
+            tmp.write(data)
+            tmp.flush()
+    except WebSocketDisconnect:
+        pass
+
+
+    finally:
+        if os.path.exists(path):
             os.remove(path)
-            print(type(result))
